# backend/go/docs/swagger.yaml
basePath: /api/v2
definitions:
...
  buyLandPayload:
    properties:
      square:
        type: integer
      x:
        type: integer
      "y":
        type: integer
    type: object
  cell:
    properties:
      _id:
        type: string
      SpendRate:
        type: number
      averageSalary:
        type: number
      cellName:
        type: string
      civilSavings:
        type: number
      crime:
        type: number
      education:
        type: number
      medicine:
        type: number
      pollution:
        type: number
      population:
        type: number
      square:
        type: integer
      surfaceImagePath:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
...
  jsonResult:
    properties:
      code:
        type: integer
      data: { }
      text:
        type: string
      values: { }
    type: object
  landLord:
    properties:
      _id:
        type: string
      square:
        type: integer
      userId:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
  logisticPayload:
    properties:
      amount:
        type: number
      buildingId:
        type: string
      fromX:
        type: integer
      fromY:
        type: integer
      resourceTypeId:
        type: integer
      toX:
        type: integer
      toY:
        type: integer
    type: object
  logisticWithData:
    properties:
      _id:
        type: string
      amount:
        type: number
      fromX:
        type: integer
      fromY:
        type: integer
      resourceType:
        $ref: '#/definitions/resourceType'
      resourceTypeId:
        type: integer
      toX:
        type: integer
      toY:
        type: integer
      userId:
        type: string
      workEnd:
        type: string
    type: object
...
  resourceType:
    properties:
      demand:
        type: number
      id:
        type: integer
      name:
        type: string
      storeGroup:
        type: string
      volume:
        description: m3
        type: number
      weight:
        description: kg
        type: number
    type: object
  resourceWithData:
    properties:
      _id:
        type: string
      amount:
        type: number
      resourceType:
        $ref: '#/definitions/resourceType'
      resourceTypeId:
        type: integer
      userId:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
...
  storage:
    properties:
      _id:
        type: string
      userId:
        type: string
      volumeMax:
        type: number
      volumeOccupied:
        type: number
      x:
        type: integer
      "y":
        type: integer
    type: object
...
host: staging.game.kube.atpstealer.com
info:
  contact: { }
  title: Game API
  version: "2.0"
paths:
...
  /map:
    get:
      description: Returns the list of all map cells
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/cell'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Return map cells
      tags:
        - map
  /map/all_land_lords:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/landLord'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Return all landowners
      tags:
        - map
  /map/buy_land:
    post:
      consumes:
        - application/json
      parameters:
        - description: Land purchase payload
          in: body
          name: buyLandPayload
          required: true
          schema:
            $ref: '#/definitions/buyLandPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  values:
                    $ref: '#/definitions/buyLandPayload'
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Buy land in cell
      tags:
        - map
  /map/cell_owners:
    get:
      consumes:
        - application/json
      parameters:
        - description: X coordinate
          in: query
          name: x
          required: true
          type: integer
        - description: Y coordinate
          in: query
          name: "y"
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/landLord'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Get the landlords in cell
      tags:
        - map
  /map/my:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/landLord'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Return user's lands
      tags:
        - map
...
  /resource/move:
    post:
      consumes:
        - application/json
      parameters:
        - description: Resource movement payload
          in: body
          name: logisticPayload
          required: true
          schema:
            $ref: '#/definitions/logisticPayload'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Initiates a resource movement
      tags:
        - logistics
  /resource/my:
    get:
      parameters:
        - description: X coordinate
          in: query
          name: x
          type: integer
        - description: Y coordinate
          in: query
          name: "y"
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/resourceWithData'
                    type: array
                type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Get user's resources
      tags:
        - resources
  /resource/my_logistics:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/logisticWithData'
                    type: array
                type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Get user's logistics tasks
      tags:
        - logistics
  /resource/types:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/resourceType'
                    type: array
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Return all resource types from database
      tags:
        - resource
...
  /storage/my:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/jsonResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/storage'
                    type: array
                type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonResult'
      summary: Return user's storages
      tags:
        - storage
...
swagger: "2.0"